Use clean architecture, MVC, Upload images to firebase storage then display in home activity.
