+++ app/src/main/AndroidManifest.xml
@@ -13,21 +13,23 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.OnlineAuctionApp">
-        <activity android:name="com.annmonstar.androidauctionapp.MyCampaigns" />
-        <activity android:name="com.annmonstar.androidauctionapp.ProductInfoActivity" />
-        <activity android:name="com.annmonstar.androidauctionapp.SettingsActivity" />
-        <activity android:name="com.annmonstar.androidauctionapp.addProductforBid" />
-        <activity android:name="com.annmonstar.androidauctionapp.HomeActivity"
+        <activity android:name="com.annmonstar.androidauctionapp.ui.MyCampaigns" />
+        <activity android:name="com.annmonstar.androidauctionapp.ui.ProductInfoActivity" />
+        <activity android:name="com.annmonstar.androidauctionapp.ui.SettingsActivity" />
+        <activity android:name="com.annmonstar.androidauctionapp.ui.addProductforBid" />
+        <activity android:name="com.annmonstar.androidauctionapp.ui.HomeActivity"
             android:exported="true">
+
+        </activity>
+        <activity android:name="com.annmonstar.androidauctionapp.ui.authentication.LoginActivity" android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name="com.annmonstar.androidauctionapp.LoginActivity" />
-        <activity android:name="com.annmonstar.androidauctionapp.RegisterActivity" />
-        <activity android:name="com.annmonstar.androidauctionapp.MainActivity" />
+        <activity android:name="com.annmonstar.androidauctionapp.ui.authentication.RegisterActivity" />
+        <activity android:name="com.annmonstar.androidauctionapp.ui.MainActivity" />
     </application>
 
 </manifest>
