+++ app/src/main/java/com/annmonstar/androidauctionapp/Adapter/AdapterClass.java
@@ -2,6 +2,7 @@ package com.annmonstar.androidauctionapp.Adapter;
 
 import android.content.Context;
 import android.content.Intent;
+import android.net.Uri;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -13,24 +14,30 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 
+import java.lang.ref.Reference;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-import com.annmonstar.androidauctionapp.HomeActivity;
+import com.annmonstar.androidauctionapp.ui.HomeActivity;
 import com.annmonstar.androidauctionapp.Models.Products;
-import com.annmonstar.androidauctionapp.ProductInfoActivity;
+import com.annmonstar.androidauctionapp.ui.ProductInfoActivity;
 import com.annmonstar.androidauctionapp.R;
+import com.google.firebase.storage.FirebaseStorage;
+import com.google.firebase.storage.StorageReference;
 
 public class AdapterClass extends RecyclerView.Adapter<AdapterClass.Viewholder> {
     List<Products>allProducts = new ArrayList<>();
     Context mContext;
+    StorageReference mStorage;
     public AdapterClass(HomeActivity homeActivity, List<Products> allProducts) {
         this.mContext =  homeActivity;
         this.allProducts = allProducts;
@@ -50,21 +57,26 @@ public class AdapterClass extends RecyclerView.Adapter<AdapterClass.Viewholder> 
         holder.pname.setText(products.getName());
         holder.pdesc.setText(products.getDescription());
         holder.pbid.setText("Ksh "+products.getBid());
+        mStorage = FirebaseStorage.getInstance().getReference();
 
         DatabaseReference reference = FirebaseDatabase.getInstance().getReference().child("Products").child(products.getName());
         reference.child("Images").addListenerForSingleValueEvent(new ValueEventListener() {
             @Override
             public void onDataChange(@NonNull DataSnapshot snapshot) {
-
-                if (!snapshot.child("image0").equals("default")){
-
+                if (!Objects.requireNonNull(snapshot.child("image0").getValue()).toString().equals("default")){
                     String url = Objects.requireNonNull(snapshot.child("image0").getValue()).toString();
-                    Glide
-                            .with(mContext)
-                            .load(url)
-                            .diskCacheStrategy(DiskCacheStrategy.DATA)
-                            .centerCrop()
-                            .into(holder.imageView);
+                    mStorage
+                            .child(url).getDownloadUrl().addOnCompleteListener(new OnCompleteListener<Uri>() {
+                                @Override
+                                public void onComplete(@NonNull Task<Uri> task) {
+                                    Glide
+                                            .with(mContext)
+                                            .load(task.getResult())
+                                            .diskCacheStrategy(DiskCacheStrategy.DATA)
+                                            .centerCrop()
+                                            .into(holder.imageView);
+                                }
+                            });
                 }
             }
 
@@ -115,6 +127,7 @@ public class AdapterClass extends RecyclerView.Adapter<AdapterClass.Viewholder> 
             pname = (TextView) itemView.findViewById(R.id.pname);
             pdesc = (TextView) itemView.findViewById(R.id.pdesc);
             pbid = (TextView) itemView.findViewById(R.id.pbid);
+
         }
     }
 }
