+++ app/src/main/java/com/annmonstar/androidauctionapp/Adapter/MyCampaignAdapter.java
@@ -24,7 +24,7 @@ import java.util.List;
 import java.util.Objects;
 
 import com.annmonstar.androidauctionapp.Models.Products;
-import com.annmonstar.androidauctionapp.ProductInfoActivity;
+import com.annmonstar.androidauctionapp.ui.ProductInfoActivity;
 import com.annmonstar.androidauctionapp.R;
