+++ app/src/main/java/com/annmonstar/androidauctionapp/ui/AuctionApp.java
@@ -1,4 +1,4 @@
-package com.annmonstar.androidauctionapp;
+package com.annmonstar.androidauctionapp.ui;
 
 import android.app.Application;
