+++ app/src/main/java/com/annmonstar/androidauctionapp/ui/SettingsActivity.java
@@ -1,4 +1,4 @@
-package com.annmonstar.androidauctionapp;
+package com.annmonstar.androidauctionapp.ui;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
@@ -14,6 +14,9 @@ import android.view.View;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
 
+import com.annmonstar.androidauctionapp.R;
+import com.annmonstar.androidauctionapp.ui.utils.PathUtils;
+import com.annmonstar.androidauctionapp.ui.utils.PathUtilvideo;
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
 import com.google.android.gms.tasks.OnCompleteListener;
