+++ app/src/main/res/layout/activity_my_campaigns.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.annmonstar.androidauctionapp.MyCampaigns">
+    tools:context="com.annmonstar.androidauctionapp.ui.MyCampaigns">
 
     <TextView
         android:id="@+id/txt"
