+++ app/src/main/res/layout/activity_settings.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.annmonstar.androidauctionapp.SettingsActivity">
+    tools:context="com.annmonstar.androidauctionapp.ui.SettingsActivity">
 
 
         <de.hdodenhof.circleimageview.CircleImageView
